+++ snapnow-api/src/main/java/kr/flab/snapnow/user/model/dto/request/account/AccountUpdateRequeest.java
@@ -1,7 +1,7 @@
-package kr.flab.snapnow.user.dto.request.account;
+package kr.flab.snapnow.user.model.dto.request.account;
 
 import java.util.Locale;
-import kr.flab.snapnow.domain.user.enums.Gender;
+import kr.flab.snapnow.domain.user.domain.enums.Gender;
 import java.time.LocalDate;
 
 import lombok.Getter;
