+++ snapnow-api/src/main/java/kr/flab/snapnow/user/model/dto/request/account/PasswordUpdateRequest.java
@@ -1,4 +1,4 @@
-package kr.flab.snapnow.user.dto.request.account;
+package kr.flab.snapnow.user.model.dto.request.account;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
