+++ snapnow-api/src/main/java/kr/flab/snapnow/user/model/dto/request/profile/ProfileImageUpdateRequest.java
@@ -1,4 +1,4 @@
-package kr.flab.snapnow.user.dto.request.profile;
+package kr.flab.snapnow.user.model.dto.request.profile;
 
 import org.springframework.web.multipart.MultipartFile;
