+++ snapnow-api/src/main/java/kr/flab/snapnow/user/model/dto/request/setting/UserSettingUpdateRequest.java
@@ -1,12 +1,12 @@
-package kr.flab.snapnow.user.dto.request.setting;
+package kr.flab.snapnow.user.model.dto.request.setting;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 
-import kr.flab.snapnow.domain.user.enums.setting.SettingCategory;
-import kr.flab.snapnow.domain.user.enums.setting.settingOption.SettingOption;
-import kr.flab.snapnow.domain.user.enums.setting.settingType.SettingType;
+import kr.flab.snapnow.domain.user.domain.enums.setting.SettingCategory;
+import kr.flab.snapnow.domain.user.domain.enums.setting.settingOption.SettingOption;
+import kr.flab.snapnow.domain.user.domain.enums.setting.settingType.SettingType;
 
 @Getter
 @NoArgsConstructor
