+++ snapnow-api/src/main/java/kr/flab/snapnow/user/model/dto/response/account/AccountResponse.java
@@ -1,4 +1,4 @@
-package kr.flab.snapnow.user.dto.response.account;
+package kr.flab.snapnow.user.model.dto.response.account;
 
 import java.time.LocalDate;
 import java.util.Locale;
@@ -6,8 +6,8 @@ import java.util.Locale;
 import lombok.Builder;
 import lombok.Getter;
 
-import kr.flab.snapnow.domain.user.enums.AuthProvider;
-import kr.flab.snapnow.domain.user.enums.Gender;
+import kr.flab.snapnow.domain.user.domain.enums.AuthProvider;
+import kr.flab.snapnow.domain.user.domain.enums.Gender;
 
 @Getter
 @Builder
