+++ snapnow-api/src/main/java/kr/flab/snapnow/user/model/dto/response/profile/UserProfileResponse.java
@@ -1,9 +1,9 @@
-package kr.flab.snapnow.user.dto.response.profile;
+package kr.flab.snapnow.user.model.dto.response.profile;
 
 import lombok.Getter;
 import lombok.Builder;
 
-import kr.flab.snapnow.domain.user.enums.FollowStatus;
+import kr.flab.snapnow.domain.user.domain.enums.FollowStatus;
 
 @Getter
 @Builder
