+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/domain/enums/setting/settingOption/SettingOption.java
